feat: more resilient startup
Looking for feedback of this prototype of a more resilient startup.
Instead of failing, initializations are retried with backoff.
Also, the error states are stored in a health variable which can
be used to report status.
The main idea is that we should know what's wrong with a server by
just looking at the status page.
If this approach is good, we'll need to adapt this to all the other
components.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>