feat(prototype): add retries during server initialization (#142)
* feat: reduce dependence on timing for tests.

There are still a few cases that are hard to avoid.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat: ticker->timertools. ticker name is too common

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat: topowrapper use exponential backoff

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat: move topowrapper to topo and use it there

I had to move topowrapper into topo due to a circular
dependency. now, topo.Store uses the WrapperConn
(renamed from topowrapper), and does not return an
error if the connection fails.
The components have been accordingly modified. For now,
I've also disabled the startup sanity checks because
we want the components to retry if they can't connect.
The components now stay up even if etcd is not
available.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat: more resilient startup

Looking for feedback of this prototype of a more resilient startup.
Instead of failing, initializations are retried with backoff.
Also, the error states are stored in a health variable which can
be used to report status.
The main idea is that we should know what's wrong with a server by
just looking at the status page.
If this approach is good, we'll need to adapt this to all the other
components.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): index page with status

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): add ready handler

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): multipooler retries on errors

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): multiorch retries on errors

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): remove checkCells

The next call after this gets a cell connection,
which can report the same error.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): simplified initialization

Refactored topo registrations into a common topopublish
framework. Also removed the need for getting hostname.
Servenv obtains it upfront, and crashes if it can't.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): refactor multiadmin

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): Change all status to index pages

Also speed up local provisioner.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): fix wrapperconn

Add a more robust way to prevent multiple simultaneous
retries. Also, add a few more connection error cases.
Also fix a flaky test.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): Fix a retry bug in topowrapper

Found another bug: an retry would always reset the
retry flag, which will cause a subsequent retry to
go through. Added a test for it to prevent this in
the future.

Also documented the known harmless race in code.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): tests for topopublish

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): auto-refresh html pages

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

* feat(polish): Publish->Register

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>

---------

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>