fix(test): replace counter-based port allocation with OS allocation (#260)
* fix(test): replace counter-based port allocation with OS allocation

Replaces the counter-based port allocation strategy with OS-based
allocation using net.Listen(":0") to eliminate port conflicts in
parallel test execution.

Changes:
- Replace GetNextPort()/GetNextEtcd2Port() with GetFreePort(t)
- Use OS to pick free ports instead of PID-based bucketing
- Add optional PeerPort to EtcdConfig for explicit etcd peer ports
- Update all test files to use new GetFreePort(t) helper
- Add ClientPort/PeerPort fields to etcdtopo.EtcdOptions

The goal here is to reduce test flakes. I'm seeing flakes related to
ports in CI and one possible explanation is that the previous method
of allocating that used a counter might have a higher chance of collisions
if multiple tests are running in parallel.

So, I'm not sure if this will decrease the rate of flakes but maybe it could?\

Even if it doesn't, centralizing how we allocate ports for tests seems like it'll
give us more flexibility to switch back to a more randomized approach than the previous
code.

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>

* Remove GenerateRandomPort()

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>

* Randomize all ports used by cluster_test and multigateway_pg_test

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>

* Fix merge issues

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>

* Set required args

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>

---------

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>