perf(multipooler): load backup_location once
This change adds the loading of backup_location to the startup of
MultiPoolerManager. This avoids having to read backup_location from the
topo on every backup/restore operation.

The change itself is small and straightforward. The bulk of the lines
changed for this PR fix tests that were broken because those tests did
not put the database in the topo.

Signed-off-by: Cuong Do <cuongdo@users.noreply.github.com>