fix(backup): replace regex w/ annotation matching
Backup was relying on fragile text matching to find the backup ID for
the backup that was just completed. Now, we add the multipooler ID and
current timestamp as annotations to the backup. Then, we use the
pgbackrest info command with JSON output to find the single backup that
matches those annotations.

Signed-off-by: Cuong Do <cuongdo@users.noreply.github.com>