chore: add dependabot cooldown
The newest zizmor linter flags a missing "cooldown"

Cooldown causes prevents dependabot from upgrading to dependencies that are too recent, which is good for supply chain safety since many hacks are short-lived. A 7-day cooldown means a supply chain hack would have to survive 7 days of public scrutiny before we'd upgrade to it.

More background: https://blog.yossarian.net/2025/11/21/We-should-all-be-using-dependency-cooldowns explains the motivation for having a cooldown,

Signed-off-by: David Weitzman <100598+dweitzman@users.noreply.github.com>