fix(provisioner): improve etcd health check
Previously, we were using /health, which ensures that the http port is
up, but does not ensure that etcd is ready to serve requests.

While /readyz exists, retrieving / is a more direct way of ensuring etcd
can service requests.  This change  centralizes the etcd health check
logic, which is also used in e2e tests.

Also added `multigres cluster restart` for convenience.

Signed-off-by: Cuong Do <cuongdo@users.noreply.github.com>